.local/bin - Added a script for ssh:ing to multiple machines at once.
They're all opened up in panes in tmux. You can then use C-o : setw synchronize panes to make it so that the same keys are sent to all the machines at the same time.

The best way of running the same interactive command across multiple machines that I've encuontered so far.